+++ lib/runners/workspace/git.rb
@@ -5,9 +5,16 @@ module Runners
     class BlameFailed < SystemError; end
 
     def range_git_blame_info(path_string, start_line, end_line)
+      @git_blame_cache ||= {}
+      cache_key = "#{path_string}\t#{start_line}\t#{end_line}"
+      cache_value = @git_blame_cache[cache_key]
+      return cache_value if cache_value
+
       stdout, _ = shell.capture3!("git", "blame", "-p", "-L", "#{start_line},#{end_line}", git_source.head, "--", path_string,
                                   trace_stdout: false, trace_stderr: true)
-      GitBlameInfo.parse(stdout)
+      GitBlameInfo.parse(stdout).tap do |result|
+        @git_blame_cache[cache_key] = result
+      end
     rescue Shell::ExecError => exn
       raise BlameFailed, "git-blame failed: #{exn.stderr_str}"
     end